+++ src/hypaad/config/__init__.py
@@ -1,0 +1,5 @@
+from .config import *
+from .study import *
+
+# pylint: disable=undefined-variable
+__all__ = config.__all__ + study.__all__