+++ CHANGELOG.rst
@@ -4,4 +4,4 @@ Changelog
 
 Unreleased
 ==========
-
+- Bayesian Optimization via Optuna